Roll back snap toolbox to v7, and add missing function
In order to use google, must authenticate within the session. Managing
multiple download sources and various authentication methods is tricky.

Further, a recent SNAP download has deprecated and removed a function
which is used within our graph, so we've rolled back to SNAP v7 until
the graph is updated to accomodate this.